+++ lib/budget/accounts/snapshot.ex
@@ -3,11 +3,10 @@ defmodule Budget.Accounts.Snapshot do
   import Ecto.Changeset
   alias Budget.Accounts.Snapshot
 
-
   schema "snapshots" do
-    field :account_id, :integer
-    field :after, :float
-    field :before, :float
+    field(:account_id, :integer)
+    field(:after, :decimal)
+    field(:before, :decimal)
 
     timestamps()
   end
