+++ lib/budget/goals/goal.ex
@@ -6,7 +6,7 @@ defmodule Budget.Goals.Goal do
   schema "goals" do
     field(:end_date, :date)
     field(:start_date, :date)
-    field(:target, :float)
+    field(:target, :decimal)
     field(:title, :string)
 
     timestamps()
